refactor dog transformation
